Rename project from Python Code MCP Server to Dev-Kit MCP Server and update related imports
