+++ README.md
@@ -1,16 +1,16 @@
-# Python Code MCP Server
+# Dev-Kit MCP Server
 
-[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/py-code-mcp-server)](https://pypi.org/project/py-code-mcp-server/)
-[![version](https://img.shields.io/pypi/v/py-code-mcp-server)](https://img.shields.io/pypi/v/py-code-mcp-server)
+[![PyPI - Python Version](https://img.shields.io/pypi/pyversions/dev-kit-mcp-server)](https://pypi.org/project/dev-kit-mcp-server/)
+[![version](https://img.shields.io/pypi/v/dev-kit-mcp-server)](https://img.shields.io/pypi/v/dev-kit-mcp-server)
 [![License](https://img.shields.io/:license-MIT-blue.svg)](https://opensource.org/licenses/MIT)
 ![OS](https://img.shields.io/badge/ubuntu-blue?logo=ubuntu)
 ![OS](https://img.shields.io/badge/win-blue?logo=windows)
 ![OS](https://img.shields.io/badge/mac-blue?logo=apple)
-[![Tests](https://github.com/DanielAvdar/py-code-mcp-server/actions/workflows/ci.yml/badge.svg)](https://github.com/DanielAvdar/py-code-mcp-server/actions/workflows/ci.yml)
-[![Code Checks](https://github.com/DanielAvdar/py-code-mcp-server/actions/workflows/code-checks.yml/badge.svg)](https://github.com/DanielAvdar/py-code-mcp-server/actions/workflows/code-checks.yml)
-[![codecov](https://codecov.io/gh/DanielAvdar/py-code-mcp-server/graph/badge.svg?token=N0V9KANTG2)](https://codecov.io/gh/DanielAvdar/py-code-mcp-server)
+[![Tests](https://github.com/DanielAvdar/dev-kit-mcp-server/actions/workflows/ci.yml/badge.svg)](https://github.com/DanielAvdar/dev-kit-mcp-server/actions/workflows/ci.yml)
+[![Code Checks](https://github.com/DanielAvdar/dev-kit-mcp-server/actions/workflows/code-checks.yml/badge.svg)](https://github.com/DanielAvdar/dev-kit-mcp-server/actions/workflows/code-checks.yml)
+[![codecov](https://codecov.io/gh/DanielAvdar/dev-kit-mcp-server/graph/badge.svg?token=N0V9KANTG2)](https://codecov.io/gh/DanielAvdar/dev-kit-mcp-server)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
-![Last Commit](https://img.shields.io/github/last-commit/DanielAvdar/py-code-mcp-server/main)
+![Last Commit](https://img.shields.io/github/last-commit/DanielAvdar/dev-kit-mcp-server/main)
 
 A Model Context Protocol (MCP) server for interacting with codebases.
 This package provides tools for turning any repository or code base into an MCP system.
@@ -27,7 +27,7 @@ This package provides tools for turning any repository or code base into an MCP 
 ## Installation
 
 ```bash
-pip install py-code-mcp-server
+pip install dev-kit-mcp-server
 ```
 
 ## Usage
@@ -36,13 +36,13 @@ pip install py-code-mcp-server
 
 ```bash
 # Recommended method (fastest startup)
-uv run python -m py_code.mcp_server
+uv run python -m dev_kit.mcp_server
 
 # Alternative method
-python -m py_code.mcp_server
+python -m dev_kit.mcp_server
 
 # Traditional method with optional parameters
-py-mcp-server --host 127.0.0.1 --port 8080
+dev-kit-mcp-server --host 127.0.0.1 --port 8080
 ```
 
 ### API Endpoints
@@ -73,8 +73,8 @@ print(f"Repository components found: {len(structure['result']['ast_analysis']['f
 
 ```bash
 # Clone the repository
-git clone https://github.com/DanielAvdar/py-code-mcp-server.git
-cd py-code-mcp-server
+git clone https://github.com/DanielAvdar/dev-kit-mcp-server.git
+cd dev-kit-mcp-server
 
 # Install development dependencies
 pip install -e ".[dev]"
