+++ dev_kit_mcp_server/__main__.py
@@ -1,7 +1,7 @@
 """Script to run the FastMCP server using uv run."""
 
 # Import the MCP instance from the module
-from py_code.fastmcp_server import start_server
+from dev_kit_mcp_server.fastmcp_server import start_server
 
 fastmcp = start_server()
 mcp = fastmcp
