+++ dev_kit_mcp_server/tools/remove.py
