+++ docs/source/index.rst
@@ -1,7 +1,7 @@
-.. Python Code MCP Server documentation master file
+.. Dev-Kit MCP Server documentation master file
 
-Python Code MCP Server
-======================
+Dev-Kit MCP Server
+===================
 
 A Model Context Protocol (MCP) server for Python code analysis.
 
@@ -16,7 +16,14 @@ Installation
 
 .. code-block:: bash
 
-   pip install py-code-mcp-server
+   pip install dev-kit-mcp-server
 
 Quick Start
 -----------
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   usage
+   api
