--- py_code_mcp_server/__init__.py
@@ -1,5 +1,0 @@
-"""Python Code MCP Server package for code analysis and repository navigation."""
-
-from py_code.fastmcp_server import start_server
-
-__all__ = ["start_server"]
