--- py_code_mcp_server/__main__.py
@@ -1,12 +1,0 @@
-"""Script to run the FastMCP server using uv run."""
-
-# Import the MCP instance from the module
-from py_code.fastmcp_server import start_server
-
-if __name__ == "__main__":
-    # Run the MCP server with the specified transport
-    # print("Starting FastMCP server...")
-    # fastmcp.run(
-    #     transport="stdio",
-    # )
-    start_server()
