+++ tests/test_main.py
@@ -23,15 +23,15 @@ def mock_args(temp_dir):
 
     with patch("argparse.ArgumentParser.parse_args", return_value=mock_args):
         # Now we can safely import the module
-        import py_code.__main__
+        import dev_kit_mcp_server.__main__
 
-        yield py_code.__main__
+        yield dev_kit_mcp_server.__main__
 
 
 @pytest.fixture
 def mock_fastmcp(mock_args):
     """Mock the FastMCP instance."""
-    with patch("py_code.fastmcp_server.start_server") as mock_start:
+    with patch("dev_kit_mcp_server.fastmcp_server.start_server") as mock_start:
         # Create a mock FastMCP instance
         mock_instance = MagicMock()
         mock_start.return_value = mock_instance
@@ -39,9 +39,9 @@ def mock_fastmcp(mock_args):
         # Reload the module to apply the mock
         import importlib
 
-        import py_code.__main__
+        import dev_kit_mcp_server.__main__
 
-        importlib.reload(py_code.__main__)
+        importlib.reload(dev_kit_mcp_server.__main__)
 
         yield mock_instance
 
@@ -50,20 +50,20 @@ def test_main_module_creates_server(mock_fastmcp, mock_args):
     """Test that the __main__ module creates a server instance."""
     # The mock_fastmcp fixture already reloads the module, which creates the instance
     # Just verify that the instance exists and is the same as our mock
-    import py_code.__main__
+    import dev_kit_mcp_server.__main__
 
-    assert py_code.__main__.fastmcp is mock_fastmcp
-    assert py_code.__main__.mcp is mock_fastmcp
+    assert dev_kit_mcp_server.__main__.fastmcp is mock_fastmcp
+    assert dev_kit_mcp_server.__main__.mcp is mock_fastmcp
 
 
 def test_main_module_runs_server_when_executed_directly(mock_fastmcp, mock_args):
     """Test that the __main__ module runs the server when executed directly."""
     # Instead of trying to trigger the if __name__ == "__main__" condition,
     # we'll directly test that the run method works as expected
-    import py_code.__main__
+    import dev_kit_mcp_server.__main__
 
     # Manually call the run method
-    py_code.__main__.fastmcp.run()
+    dev_kit_mcp_server.__main__.fastmcp.run()
 
     # Verify that run() was called
     mock_fastmcp.run.assert_called_once()
